Add `wait_for_log`
